requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-ckeditor==6.0.0
django-instagram==0.3.2
django-js-asset==1.2.2
httplib2==0.19.0
idna==2.10
lxml==4.6.2
Pillow==8.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.3
